% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_function.R
\name{make_blocks}
\alias{make_blocks}
\title{Title Creating Qualtrics Blocks}
\usage{
make_blocks(blocks)
}
\arguments{
\item{blocks}{Required. Blocks must be a list with each element of the list being the contents for one block. Each element of blocks should be one of the make_qualtrics functions (e.g., make_qualtrics_MC, make_qualtrics_dropdown, make_qualtrics_matrix)}
}
\value{
make_blocks outputs a vector that separates each element of the list blocks into a separate block for Qualtrics to import.
}
\description{
Title Creating Qualtrics Blocks
}
\examples{
\dontrun{ 
  x <- list("I like school", "Doing well in school is not important", "I don't need to do well in school to succeed in life")
  y <- list(c("Describes me very well", "Does not Describe me at all"), c("Agree", "Disagree"), c("Very True", "Somewhat True", "Neither True nor False", "Somewhat False", "Very False"))
  a <- c("I feel calm", "I feel happy", "I feel tired", "I feel sad")
  b <- c("Strongly Disagree", "Disagree", "Neutral", "Agree", "Strongly Agree")
  c <- "Please read each question and answer each statement honestly using the following scale."
  qualtrics <- list(make_qualtrics_mc_dropdown(x, y), make_qualtrics_matrix(a, b, c))
  qualtrics <- make_blocks(qualtrics)
}
}
