% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_function.R
\name{make_qualtrics_mc}
\alias{make_qualtrics_mc}
\title{Title Make a Multiple Choice (including Likert Scale) in Qualtrics}
\usage{
make_qualtrics_mc(questions, answer_scale, .items_per_page = 1L)
}
\arguments{
\item{questions}{Required. The questions parameter is the stem, or the questions, that you would like to present. This should be a vector, list, or column in a dataset that contains character strings.}

\item{answer_scale}{Required. The answer_scale parameter is the options that you want to give your participants for each of the questions presented. This should be a vector or a list of length one that contains all the answer choices you would like.}

\item{.items_per_page}{Not required, will default to 1. Option allowing number of items to be shown per page. This should be a whole number.}
}
\value{
This function will return a vector that is in the required format to import the question as a dropdown into qualtrics.
}
\description{
Title Make a Multiple Choice (including Likert Scale) in Qualtrics
}
\examples{
\donotrun{
x <- c("I feel calm", "I feel sad", "I feel mad", "I feel happy")
y <- rep(list(c("Yes", "No"), c("Strongly Agree", "Agree", "Neutral", "Disagree", "Strongly Disagree")), 2)
make_qualtrics_mc(x, y, .items_per_page = 3L)
}
}
