% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_function.R
\name{make_qualtrics_mc_multiSelect}
\alias{make_qualtrics_mc_multiSelect}
\title{Title Multiple Choice with Multiple Answers}
\usage{
make_qualtrics_mc_multiSelect(questions, answer_scale, .items_per_page = 1L)
}
\arguments{
\item{questions}{Required. The questions parameter is the stem, or the questions, that you would like to present. This should be a vector, list, or column in a dataset that contains character strings.}

\item{answer_scale}{Required. The answer_scale parameter is the options that you want to give your participants for each of the questions presented. This should be a vector or a list of length one that contains all the answer choices you would like.}

\item{.items_per_page}{Not required, will default to 1. Option allowing number of items to be shown per page. This should be a whole number.}
}
\value{
This function will return a vector that is in the required format to import the question stems and options as a multiple choice questions that participants can select more than one option into qualtrics.
}
\description{
Title Multiple Choice with Multiple Answers
}
\examples{
\dontrun{ 
x <- list("I like school", "Doing well in school is not important", "I don't need to do well in school to succeed in life")
y <- list(c("Describes me very well", "Does not Describe me at all"), c("Agree", "Disagree"), c("Very True", "Somewhat True", "Neither True nor False", "Somewhat False", "Very False"))
make_qualtrics_mc_multiSelect(x, y, .item_per_page = 2L)
}
}
