% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_function.R
\name{output_file}
\alias{output_file}
\title{Title Outputting Qualtrics Survey to txt File}
\usage{
output_file(blocks, set_dir = getwd(), out_file = "qualtrics_import.txt")
}
\arguments{
\item{blocks}{Required. The blocks argument is the saved variable from the make_blocks function that includes all the desired blocks for the survey}

\item{set_dir}{Optional. set_dir is the file path that you want to file to save in. Default is current working directory}
}
\value{
output_file will save blocks to a .txt file in a set directory so that user can upload the .txt file to Qualtrics
}
\description{
Title Outputting Qualtrics Survey to txt File
}
\examples{
\dontrun{ 
x <- list("I like school", "Doing well in school is not important", "I don't need to do well in school to succeed in life")
y <- list(c("Describes me very well", "Does not Describe me at all"), c("Agree", "Disagree"), c("Very True", "Somewhat True", "Neither True nor False", "Somewhat False", "Very False"))
a <- c("I feel calm", "I feel happy", "I feel tired", "I feel sad")
b <- c("Strongly Disagree", "Disagree", "Neutral", "Agree", "Strongly Agree")
c <- "Please read each question and answer each statement honestly using the following scale."
qualtrics <- list(make_qualtrics_mc_dropdown(x, y), make_qualtrics_matrix(a, b, c))
qualtrics <- make_blocks(qualtrics)
output_file(qualtrics)
}
}
