% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_break_function.R
\name{page_break}
\alias{page_break}
\title{Page Break helper function}
\usage{
page_break(.x, .items_per_page = 1L)
}
\arguments{
\item{.x}{Not Required. Will be assigned within function}

\item{.items_per_page}{Not required. Will default as 1 item per page but can
changed if specified}
}
\value{
This is a helper function that will add \code{[[PageBreak]]} at a specified
interval between items
}
\description{
Page Break helper function
}
\examples{
\donotrun{ 
x <- list("I like school", "Doing well in school is not important", "I don't need to do well in school to succeed in life")
y <- list(c("Describes me very well", "Does not Describe me at all"), c("Agree", "Disagree"), c("Very True", "Somewhat True", "Neither True nor False", "Somewhat False", "Very False"))
make_qualtrics_mc_dropdown(x, y, .items_per_page = 3L)
}
}
